+++ pywren_ibm_cloud/storage/storage.py
@@ -4,7 +4,7 @@ import pickle
 import logging
 import importlib
 from pywren_ibm_cloud.version import __version__
-from pywren_ibm_cloud.config import CONFIG_DIR
+from pywren_ibm_cloud.config import CACHE_DIR
 from pywren_ibm_cloud.utils import is_remote_cluster
 from pywren_ibm_cloud.storage.utils import create_status_key, create_output_key, \
     status_key_suffix, CloudObject, StorageNoSuchKeyError
@@ -180,9 +180,9 @@ class InternalStorage:
         :return: runtime metadata
         """
         path = ['runtimes', __version__,  key+".meta.json"]
-        filename_local_path = os.path.join(CONFIG_DIR, *path)
+        filename_local_path = os.path.join(CACHE_DIR, *path)
 
-        if os.path.exists(filename_local_path) and not is_remote_cluster:
+        if os.path.exists(filename_local_path) and not is_remote_cluster():
             logger.debug("Runtime metadata found in local cache")
             with open(filename_local_path, "r") as f:
                 runtime_meta = json.loads(f.read())
@@ -215,8 +215,8 @@ class InternalStorage:
         logger.debug("Uploading Runtime metadata to: /{}/{}".format(self.bucket, obj_key))
         self.storage_handler.put_object(self.bucket, obj_key, json.dumps(runtime_meta))
 
-        if not is_remote_cluster:
-            filename_local_path = os.path.join(CONFIG_DIR, *path)
+        if not is_remote_cluster():
+            filename_local_path = os.path.join(CACHE_DIR, *path)
             logger.debug("Saving runtime metadata into local cache: {}".format(filename_local_path))
 
             if not os.path.exists(os.path.dirname(filename_local_path)):
@@ -233,7 +233,7 @@ class InternalStorage:
         """
         path = ['runtimes', __version__,  key+".meta.json"]
         obj_key = '/'.join(path).replace('\\', '/')
-        filename_local_path = os.path.join(CONFIG_DIR, *path)
+        filename_local_path = os.path.join(CACHE_DIR, *path)
         if os.path.exists(filename_local_path):
             os.remove(filename_local_path)
         self.storage_handler.delete_object(self.bucket, obj_key)